Move field components to separate files
