+++ webapp/src/components/Settings/utils.ts
@@ -1,0 +1,7 @@
+export type FieldProps<T> = { value: T; onChange?: (newValue: T) => void };
+
+export const FIELD_COMMON_PROPS = {
+  size: "small",
+  variant: "standard",
+  InputLabelProps: { shrink: true },
+} as const;
